extract utility functions to core/helpers.js
