+++ public/app.js
@@ -1,3 +1,5 @@
+// Extracted helpers
+import { formatTitle, $, h, fmtSize, fmtDate, formatTimestamp } from '/core/helpers.js';
 // --- tiny router (hash-based) ---
 const routes = {
   "": renderHome,
@@ -241,12 +243,7 @@ async function renderWatch() {
   }
 }
 
-// Utility: format seconds as mm:ss
-function formatTimestamp(ts) {
-  const min = Math.floor(ts / 60);
-  const sec = ts % 60;
-  return `${min}:${sec.toString().padStart(2, "0")}`;
-}
+// Utility moved to /core/helpers.js: formatTimestamp
 
 // --- Moments Page ---
 async function renderMoments() {
@@ -408,34 +405,7 @@ async function toggleFav(e, item) {
   onRoute();
 }
 
-// --- utils ---
-// Format video title for display: replace underscores with spaces and remove extension
-function formatTitle(name) {
-  let base = name.replace(/\.[^/.]+$/, "");
-  return base.replace(/_/g, " ");
-}
-function $(sel, root=document){ return root.querySelector(sel); }
-function h(tag, attrs={}, ...children){
-  const el = document.createElement(tag);
-  Object.entries(attrs).forEach(([k,v])=>{
-    if (k === "class") el.className = v;
-    else if (k.startsWith("on") && typeof v === "function") el.addEventListener(k.slice(2), v);
-    else if (v !== false && v != null) el.setAttribute(k, v === true ? "" : v);
-  });
-  children.flat().forEach(c => {
-    if (c == null) return;
-    el.append(c.nodeType ? c : document.createTextNode(c));
-  });
-  return el;
-}
-function fmtSize(bytes){
-  if (!bytes && bytes !== 0) return "";
-  const u = ["B","KB","MB","GB","TB"];
-  let i = 0, n = bytes;
-  while (n >= 1024 && i < u.length-1) { n/=1024; i++; }
-  return `${n.toFixed(1)} ${u[i]}`;
-}
-function fmtDate(ms){ return new Date(ms).toLocaleString(); }
+// --- utils --- (moved to /core/helpers.js)
 
 async function api(url){
   const r = await fetch(url);
